backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "interviewready-backend"
version = "1.0.0"
description = "Production-Ready Multi-Agent AI Backend with Gemini"
authors = [
    {name = "InterviewReady Team", email = "team@interviewready.com"},
]
dependencies = [
    "fastapi[standard]>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.4.0",
    "pydantic-settings>=2.0.0",
    "sqlalchemy>=2.0.0",
    "firebase-admin>=6.0.0",
    "google-generativeai>=0.3.0",
    "google-genai==1.67.0",
    "langchain-google-genai>=0.0.5",
    "langgraph>=0.1.0",
    "python-dotenv>=1.0.0",
    "pypdf>=6.9.1",
    "langfuse>=3.0.0",
    "llm-guard>=0.3.0",
    "datafog[nlp]>=4.3.0",
    "pip",
    "langchain>=1.2.13",
    "websockets>=13.0",
    "slowapi>=0.1.9",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "pre-commit>=3.4.0",
]

[tool.hatch.build]
include = ["app/mock_responses.json"]

[tool.hatch.build.targets.wheel]
packages = ["app"]
include = ["app/mock_responses.json"]

[tool.hatch.build.targets.sdist]
include = ["app/mock_responses.json"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.black]
line-length = 88
target-version = ['py311']

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`) codes by default.
# Unlike Flake8, Ruff doesn't enable pycodestyle warnings (`W`) or
# McCabe complexity (`C901`) by default.
select = ["E4", "E7", "E9", "F", "I"]
ignore = []

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.15.6",
]
